poetry-demo/pyproject.toml
[tool.poetry]
name = "poetry-demo"
version = "0.1.0"
description = "Python virtual environment playground using poetry"
authors = ["mikejmnz24 <4935341+mikejmz24@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.31.0"
pytest = "^8.1.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
